+++ web/src/components/organisms/Plant/Plant/index.ts
@@ -1,0 +1,1 @@
+export * from './Plant'
