+++ web/src/components/organisms/Plant/PlantsList/index.ts
@@ -1,0 +1,1 @@
+export * from './PlantsList'
