+++ web/src/components/organisms/Plant/index.ts
@@ -1,0 +1,2 @@
+export * from './Plant'
+export * from './PlantsList'
