+++ web/src/components/organisms/index.ts
@@ -1,0 +1,1 @@
+export * from './Plant'
